test(config): 将 UsernameAuthenticatorSerializerTest 迁移到 SerializersKtTest.
将测试类移动到对应的文件中.
