+++ scalabot-meta/src/test/kotlin/UsernameAuthenticatorTest.kt
@@ -1,11 +1,8 @@
 package net.lamgc.scalabot.config
 
-import com.google.gson.*
-import net.lamgc.scalabot.config.serializer.UsernameAuthenticatorSerializer
-import org.junit.jupiter.api.assertThrows
 import kotlin.test.*
 
-class UsernameAuthenticatorTest {
+internal class UsernameAuthenticatorTest {
 
     @Test
     fun checkCredentialsTest() {
@@ -35,81 +32,6 @@ class UsernameAuthenticatorTest {
         assertNotEquals(authenticator, UsernameAuthenticator("falseUser", "falsePassword"))
         assertNotEquals(authenticator.hashCode(), UsernameAuthenticator("falseUser", "falsePassword").hashCode())
         assertFalse(authenticator.equals(null))
-    }
-
-}
-
-class UsernameAuthenticatorSerializerTest {
-
-    @Test
-    fun serializeTest() {
-        val authenticator = UsernameAuthenticator("testUser", "testPassword")
-        val jsonElement = UsernameAuthenticatorSerializer.serialize(authenticator, null, null)
-        assertTrue(jsonElement.isJsonObject)
-        val jsonObject = jsonElement.asJsonObject
-        assertEquals("testUser", jsonObject["username"]?.asString)
-        assertEquals("testPassword", jsonObject["password"]?.asString)
-    }
-
-    @Test
-    fun deserializeTest() {
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonArray(), null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonPrimitive(""), null, null)
-        }
-        assertNull(UsernameAuthenticatorSerializer.deserialize(JsonNull.INSTANCE, null, null))
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                addProperty("username", "testUser")
-            }, null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                addProperty("username", "testUser")
-                add("password", JsonArray())
-            }, null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                addProperty("password", "testPassword")
-            }, null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                add("username", JsonArray())
-                addProperty("password", "testPassword")
-            }, null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                addProperty("username", "")
-                addProperty("password", "")
-            }, null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                addProperty("username", "testUser")
-                addProperty("password", "")
-            }, null, null)
-        }
-        assertThrows<JsonParseException> {
-            UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-                addProperty("username", "")
-                addProperty("password", "testPassword")
-            }, null, null)
-        }
-
-        val authenticator = UsernameAuthenticatorSerializer.deserialize(JsonObject().apply {
-            addProperty("username", "testUser")
-            addProperty("password", "testPassword")
-        }, null, null)
-        assertNotNull(authenticator)
-
-        assertTrue(authenticator.checkCredentials("testUser", "testPassword"))
-        assertFalse(authenticator.checkCredentials("falseUser", "testPassword"))
-        assertFalse(authenticator.checkCredentials("testUser", "falsePassword"))
     }
 
 }
